{{cookiecutter.project_name}}/requirements/base.txt
# ASGI
starlette==0.13.3

# deploy
uvicorn==0.11.3
gunicorn==20.0.4

# types & validator
pydantic==1.5.1

# GraphQL
python-gql==0.0.14
starlette-graphql==0.1.2
async-dataloader==0.0.1

# database
alembic==1.4.2
databases==0.3.2
SQLAlchemy==1.3.15
psycopg2-binary==2.8.4

# redis
aioredis==1.3.1

# file storage
minio==5.0.8
python-multipart==0.0.5

# http requests
httpx==0.12.1

# task schedule
celery==4.4.5
